requirements.txt
Django>=3.0,<4.0
djangorestframework>1.0
celery>3
kombu>5